% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggrid.R
\name{aggrid_output_handler}
\alias{aggrid_output_handler}
\title{Data Grid Output Handler}
\usage{
aggrid_output_handler(data, ...)
}
\arguments{
\item{data}{the JSON data}

\item{...}{other unused parameters}
}
\value{
a data.frame
}
\description{
Transform exported data into data.frame. Convert null and Nan to 0 or dataFrame will not like it
}
