% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggrid.R
\name{aggridUI}
\alias{aggridUI}
\title{Prepare UI tag and Dependency for Grid.}
\usage{
aggridUI(id = "grid-container", width = "100\%", height = "400px")
}
\arguments{
\item{id}{This is used exactly once, in the JavaScript creating the grid}

\item{width}{The width of the input, e.g. \verb{400px}, or \verb{100\%}; see \code{htmltools::validateCssUnit()}.}

\item{height}{The height of the input, e.g. \verb{400px}, or \verb{100\%}; see \code{htmltools::validateCssUnit()}.}
}
\value{
Placeholder HTML for grid.
}
\description{
Prepare UI tag and Dependency for Grid.
}
