% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datagridxl.R
\name{datagridxl}
\alias{datagridxl}
\title{Prepare Data and Options for DataGridXL}
\usage{
datagridxl(data, options = list())
}
\arguments{
\item{data}{data.frame to show}

\item{options}{list of options according to DataGridXL documentation}
}
\value{
list
}
\description{
See sample app in package for example: \code{system.file("example/nonreactive/app.R", package = "datagridxlr")}
}
