% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datagridxl.R
\name{datagridxl_output_handler}
\alias{datagridxl_output_handler}
\title{Data Grid Output Handler}
\usage{
datagridxl_output_handler(data, ...)
}
\arguments{
\item{data}{the JSON data}

\item{...}{other unused parameters}
}
\value{
a data.frame
}
\description{
Transform exported data into data.frame.
}
